+++ src/test/java/roomescape/service/ThemeServiceTest.java
@@ -8,7 +8,7 @@ import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.autoconfigure.jdbc.JdbcTest;
 import org.springframework.jdbc.core.JdbcTemplate;
-import roomescape.dao.ThemeRepository;
+import roomescape.repository.ThemeRepository;
 import roomescape.domain.Theme;
 
 import java.util.List;
